requirements_dataprocess.txt
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
math-verify[antlr4_9_3]
numpy
pandas
peft
pyarrow>=15.0.0
pybind11
pylatexenc
ray[default]
tensordict<0.6
torchdata
transformers
vllm==0.8.2
wandb
openai
streamlit
git+https://github.com/illuin-tech/colpali
transformers==4.50.3
llama-cloud==0.1.5
llama-index==0.12.0
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.1
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-openai==0.3.0
llama-index-indices-managed-llama-cloud==0.6.2
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.1
llama-index-multi-modal-llms-openai==0.3.0
llama-index-program-openai==0.3.0
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.0
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.14
pdf2image